A few minor fixes + `V2` naming refactor for `modeling` subpackage (#816)
* make sure to use mpl agg backend

* allow cache to be created on demand

* fix protein prep naming for V2 and unused code

* refactor testing

* finish refactoring testing

* add in print silencing

* finish fixing tests

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix imports

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix issues with context manager

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix assignment

---------
