+++ asapdiscovery-data/asapdiscovery/data/logging.py
@@ -1,5 +1,6 @@
 import logging
 import os
+import sys
 from typing import Optional
 
 from rich.console import Console
@@ -66,3 +67,14 @@ seem to work nicely and the log will be difficult to read and possibly missing i
     oechem.OEThrow.SetOutputStream(errfs)
 
 """
+
+
+class HiddenPrint:
+    def __enter__(self):
+        self._original_stdout = sys.stdout
+        sys.stdout = open(os.devnull, "w")
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        sys.stdout.close()
+        sys.stdout = self._original_stdout
