+++ asapdiscovery-docking/asapdiscovery/docking/workflows/large_scale_docking.py
@@ -48,7 +48,7 @@ from asapdiscovery.docking.scorer_v2 import (
 )
 from asapdiscovery.docking.workflows.workflows import PosteraDockingWorkflowInputs
 from asapdiscovery.ml.models import ASAPMLModelRegistry
-from asapdiscovery.modeling.protein_prep_v2 import ProteinPrepper
+from asapdiscovery.modeling.protein_prep import ProteinPrepper
 from distributed import Client
 from pydantic import Field, PositiveInt, validator
